# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: HorneroDB API
  description: |
    CRM-style database-as-a-service API.
    Supports dynamic tables, columns, records, role-based permissions (Dataverse-style),
    SSO via OIDC, and API key authentication.
  version: "1.0.0"
  contact:
    name: HorneroDB
  license:
    name: MIT
servers:
  - url: http://localhost:8090/api/v1
    description: Local development

# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
#  SECURITY
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API Key prefixed with `Bearer `"

  # ── Schemas ──────────────────────
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Message:
      type: object
      properties:
        message:
          type: string

    # ── Workspace ──
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        owner_id:
          type: string
          format: uuid
        settings:
          type: object
        created_at:
          type: string
          format: date-time
    WorkspaceCreate:
      type: object
      required: [name, slug, owner_id]
      properties:
        name:
          type: string
        slug:
          type: string
        owner_id:
          type: string
          format: uuid

    # ── Table ──
...
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TableSlug:
      name: table_slug
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC callback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect with JWT token
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List all workspaces
      description: "Requires User Session (OIDC). Instance-level API Keys (WorkspaceID = Nil) are allowed, standard API Keys are blocked."
      responses:
        "200":
          description: Array of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      tags: [Workspaces]
      summary: Create a workspace
      description: |
        Requires User Session (OIDC) or an Instance-Level API Key.
        Also creates default admin and user roles, assigns admin role to owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /workspaces/{workspace_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspaces]
      summary: Get a workspace
      description: "Requires User Session or API Key assigned to the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Workspaces]
      summary: Update a workspace
      description: "Requires 'admin' role within the workspace."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                settings:
                  type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Workspaces]
      summary: Delete a workspace
      description: "Requires 'admin' role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  TABLES
  # ══════════════════════════════════
  /workspaces/{workspace_id}/tables:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Tables]
      summary: List tables in workspace
      description: "Requires any assigned role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      tags: [Tables]
      summary: Create a table
      description: |
        Requires 'admin' role within the workspace.
        Creates metadata entry and physical PostgreSQL table with id, created_by, created_at, updated_at.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableCreate"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /workspaces/{workspace_id}/tables/{table_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TableId"
    get:
      tags: [Tables]
      summary: Get a table
      description: "Requires any assigned role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    put:
      tags: [Tables]
      summary: Update table metadata
      description: "Requires 'admin' role within the workspace."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Tables]
      summary: Delete a table
      description: |
        Requires 'admin' role within the workspace.
        Drops the physical PostgreSQL table and removes metadata.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  COLUMNS
  # ══════════════════════════════════
  /workspaces/{workspace_id}/tables/{table_id}/columns:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TableId"
    get:
      tags: [Columns]
      summary: List columns of a table
      description: "Requires any assigned role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Column"
    post:
      tags: [Columns]
      summary: Create a column
      description: |
        Requires 'admin' role within the workspace.
        Adds column metadata and ALTER TABLEs the physical table.
        Supported field_type values and their SQL mappings:

        | field_type | PostgreSQL type |
        |------------|----------------|
        | text | VARCHAR(255) |
        | long_text | TEXT |
        | number | NUMERIC |
        | integer | INTEGER |
        | float | DOUBLE PRECISION |
        | boolean | BOOLEAN |
        | date | DATE |
        | datetime | TIMESTAMPTZ |
        | email | VARCHAR(255) |
        | url | VARCHAR(500) |
        | select | VARCHAR(100) |
        | relation | UUID |
        | json | JSONB |
        | attachment | JSONB |
        | autonumber | VARCHAR(255) (sequential ID e.g. PED-001, meta: {prefix: "PED-", digits: 3, current_value: 1}) |
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColumnCreate"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Column"
  /workspaces/{workspace_id}/tables/{table_id}/columns/{column_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TableId"
      - $ref: "#/components/parameters/ColumnId"
    put:
      tags: [Columns]
      summary: Update column definition and alter physical table
      description: |
        Requires 'admin' role within the workspace.
        Updates column metadata and performs physical ALTER TABLE (renaming column or altering column SQL data type) if slug or field_type changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                field_type:
                  type: string
                  enum: [text, long_text, number, integer, float, boolean, date, datetime, email, url, select, relation, json, attachment]
                meta:
                  type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Columns]
      summary: Delete a column
      description: |
        Requires 'admin' role within the workspace.
        Drops the physical column and removes metadata.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  DATA (RECORDS)
  # ══════════════════════════════════
  /workspaces/{workspace_id}/data/{table_slug}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TableSlug"
    get:
      tags: [Records]
      summary: List records
      description: "Requires 'read' permission (all/own) on the table. AccessOwn filters by created_by."
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordList"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Records]
      summary: Create a record
      description: "Requires 'create' permission (all/own) on the table."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Record"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /workspaces/{workspace_id}/data/{table_slug}/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/TableSlug"
      - $ref: "#/components/parameters/RecordId"
    get:
      tags: [Records]
      summary: Get a single record
      description: "Requires 'read' permission (all/own) on the table."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Records]
      summary: Update a record
      description: "Requires 'update' permission (all/own) on the table."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Record"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Records]
      summary: Delete a record
      description: "Requires 'delete' permission (all/own) on the table."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  ROLES (Dataverse-style)
  # ══════════════════════════════════
  /workspaces/{workspace_id}/roles:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Roles]
      summary: List roles
      description: "Requires any assigned role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      tags: [Roles]
      summary: Create a role
      description: "Requires 'admin' role within the workspace."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /workspaces/{workspace_id}/roles/{role_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/RoleId"
    get:
      tags: [Roles]
      summary: Get a role
      description: "Requires any assigned role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    put:
      tags: [Roles]
      summary: Update a role
      description: "Requires 'admin' role within the workspace."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                permissions:
                  type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags: [Roles]
      summary: Delete a role
      description: "Requires 'admin' role within the workspace. Cannot delete default roles."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  USER ROLES
  # ══════════════════════════════════
  /workspaces/{workspace_id}/users:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [UserRoles]
      summary: List user-role assignments
      description: "Requires 'admin' role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRole"
  /workspaces/{workspace_id}/users/{user_id}/role:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [UserRoles]
      summary: Assign role to user
      description: "Requires 'admin' role within the workspace."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_id]
              properties:
                role_id:
                  type: string
                  format: uuid
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"
        "200":
          description: Updated existing assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"
    delete:
      tags: [UserRoles]
      summary: Remove role from user
      description: "Requires 'admin' role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ══════════════════════════════════
  #  PERMISSIONS (legacy)
  # ══════════════════════════════════
  /workspaces/{workspace_id}/permissions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Permissions]
      summary: List permissions
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
    post:
      tags: [Permissions]
      summary: Create a permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Permission"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
...
  /workspaces/{workspace_id}/keys/{key_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/KeyId"
    delete:
      tags: [APIKeys]
      summary: Delete an API key
      description: "Requires 'admin' role within the workspace."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"